requirements.txt
loguru
apscheduler
flask
flask-cors
flask-wtf
sqlalchemy
sqlalchemy_serializer
pymysql
yahooquery
backtrader
pyyaml
requests
bs4
retrying
# pyfolio==0.9.2
# pymysql==1.1.0

certifi
gunicorn
gevent
# wtforms
# jwt